Put Move in Commands
