+++ LibGit2Sharp.Tests/FileHistoryFixture.cs
@@ -152,7 +152,7 @@ namespace LibGit2Sharp.Tests
 
                 // Move the first file to a new directory.
                 var newPath1 = Path.Combine(SubFolderPath1, path1);
-                repo.Move(path1, newPath1);
+                Commands.Move(repo, path1, newPath1);
                 var commit3 = repo.Commit("Moved " + path1 + " to " + newPath1,
                     Constants.Signature, Constants.Signature);
